==> notebooks/05_Tests2_Exams.ipynb <==
import json
import re
import pandas as pd
import random
from dotenv import load_dotenv
from llama_index.core import (
    Document,
    VectorStoreIndex,
    StorageContext,
    load_index_from_storage,
    get_response_synthesizer,
)
from llama_index.core.node_parser import SimpleNodeParser
from llama_index.core.query_engine import RetrieverQueryEngine
from llama_index.core.evaluation import RetrieverEvaluator
from llama_index.core.prompts import PromptTemplate
from llama_index.core.schema import QueryBundle
from llama_index.core.settings import Settings
from llama_index.llms.openai import OpenAI
from llama_index.embeddings.openai import OpenAIEmbedding

# Carga del texto limpio
data_path = "../data/plain_text/plain_text.txt"
with open(data_path, "r", encoding = "utf-8") as f:
    content = f.read()

# Generar el document
pdf_doc = Document(text=content.strip())

# Carga de las preguntas
qa_path = "../data/exams_test2/clean_questions/questions_all_tests.json"
with open(qa_path, "r", encoding="utf-8") as f:
    questions = json.load(f)

# Export boolean
export_csv = False

# Key
load_dotenv()
# Set embedding model
Settings.embed_model = OpenAIEmbedding(model="text-embedding-3-small")

# Primera Prueba

# LLM Model
system_prompt = """
Eres un experto en Apache Spark y PySpark, y estás respondiendo preguntas tipo test usando exclusivamente la información del contexto proporcionado.

No puedes utilizar ningún conocimiento previo ni información externa al contexto.

Tu tarea es:
1. Leer el contexto cuidadosamente.
2. Leer la pregunta y las opciones tipo test.
3. Seleccionar la opción correcta **solo si** está claramente justificada por el contenido del contexto.
4. Justificar tu elección **citando explícitamente fragmentos relevantes del contexto**.

Si no puedes justificar una respuesta a partir del contexto, responde exactamente "No he encontrado información al respecto." y no devuelvas ninguna opción como respuesta.
"""
llm = OpenAI(
    model="gpt-4o-mini",
    temperature=0,
    system_prompt=system_prompt
)

Settings.llm = llm
# Set embedding model
Settings.embed_model = OpenAIEmbedding(model="text-embedding-3-small")

# Ruta al índice optimizado
storage_path = "../data/index_storage/"
storage_context = StorageContext.from_defaults(persist_dir=storage_path)
index = load_index_from_storage(storage_context)

# Configurar el retriever
retriever = index.as_retriever(similarity_top_k=5)

# Función auxiliar para formatear las opciones:
def format_options(options_dict):
    clean_options = {}
    for key in options_dict:
        # Elimina letras errantes como '\nB' o '\nC' dentro del texto
        value = options_dict[key]
        value = re.sub(r'\n[A-E]\n?', '', value).strip()
        clean_options[key] = value
    return "\n".join([f"{k}. {v}" for k, v in clean_options.items()])

# Función para lanzar la query:
def ask_question_with_options(question_obj, 
                              #query_str_template, 
                              retriever = retriever, response_mode= "compact"):
    question_text = question_obj["question"]
    options_text = format_options(question_obj["options"])

    #qa_prompt = PromptTemplate(query_str_template)
    #Settings.text_qa_template = qa_prompt

    query_str = f"""
    {question_text}

    Opciones: 
    {options_text}
        
    IMPORTANTE:
        - Usa solo información del contexto.
        - No respondas con conocimientos externos.
        - Si no puedes justificar la respuesta con fragmentos del contexto, no devuelvas ninguna opción.
        - Si puedes justificar la respuesta con fragmentos del contexto, justificala
    """
    
    # Usar QueryBundle para pasar la información
    query_bundle = QueryBundle(
        query_str = query_str,
        custom_embedding_strs = [question_text, options_text]
    )

    response_synthesizer = get_response_synthesizer(response_mode=response_mode)
    
    query_engine = RetrieverQueryEngine(
        retriever=retriever,
        response_synthesizer=response_synthesizer
    )
    
    response = query_engine.query(query_bundle)
    predicted_answer = response.response.strip()

    return {
        "question_id": question_obj["question_id"],
        "question" : question_obj["question"],
        "options": format_options(question_obj["options"]),
        "predicted_answer": predicted_answer,
        "correct_answer": question_obj["correct_answer"],
        "is_correct": predicted_answer == question_obj["correct_answer"],
        "response_obj": response  # Para inspeccionar source_nodes si se desea
    }

# Validar la repsuesta:
def display_question_result(result):
    print("ID Pregunta:", result["question_id"])
    print("\nPregunta:\n", result["question"])
    print("\nOpciones:\n", result["options"])
    print("\nAcertó?:", result["is_correct"])
    print("\nRespuesta generada por el modelo:\n", result["response_obj"])
    print("\nRespuesta correcta:", result["correct_answer"])
    if "response_obj" in result:
        print(f"\n🔎 Chunks usados para responder la pregunta {result['question_id']}:\n")
        for i, node in enumerate(result["response_obj"].source_nodes):
            print(f"--- Chunk #{i+1} | Score: {node.score:.4f} ---")
            print(node.node.get_content().strip())
            print()

question =[q for q in questions if q["question_id"] == "Test_2_39"][0]
test_result_2_39 = ask_question_with_options(question, response_mode = "compact")
# Validation
display_question_result(test_result_2_39)

question =[q for q in questions if q["question_id"] == "Test_1_27"][0]
test_result_2_39 = ask_question_with_options(question, response_mode = "compact")
# Validation
display_question_result(test_result_2_39)

question =[q for q in questions if q["question_id"] == "Test_2_54"][0]
test_result_2_39 = ask_question_with_options(question, response_mode = "compact")
# Validation
display_question_result(test_result_2_39)

# Test_2_18, Test_1_9, Test_2_12, Test_1_53, Test_1_41

ids = ["Test_2_18", "Test_1_9", "Test_2_12", "Test_1_53", "Test_1_41"]

for id in ids: 
    question =[q for q in questions if q["question_id"] == id][0]
    #result_random = ask_question_with_options(question, response_mode = "refine")
    result_random = ask_question_with_options(question, response_mode = "tree_summarize")
    display_question_result(result_random)
    print("="*20)

ids = ["Test_1_53", "Test_1_17", "Test_2_28", "Test_2_31", "Test_2_29"]

for id in ids: 
    question =[q for q in questions if q["question_id"] == id][0]
    #result_random = ask_question_with_options(question, response_mode = "refine")
    result_random = ask_question_with_options(question, response_mode = "tree_summarize")
    display_question_result(result_random)
    print("="*20)

random_questions = random.sample(questions, k=5)
for q in random_questions: 
    result_random = ask_question_with_options(q, response_mode = "refine")
    # result_random = ask_question_with_options(q, response_mode = "tree_summarize")
    display_question_result(result_random)
    print("="*20)

result_random = ask_question_with_options(random.choice(questions), response_mode = "refine")
display_question_result(result_random)

# System Prompt menos restrictivo - Segunda Prueba

* Este prompt permitiría razonar desde el contenido si no está claramente la respuesta en el contexto.

system_prompt = """
Eres un experto en Apache Spark y PySpark, y estás respondiendo preguntas tipo test usando exclusivamente la información del contexto proporcionado.

No puedes utilizar ningún conocimiento previo ni información externa al contexto.

Tu tarea es:
1. Leer el contexto cuidadosamente.
2. Leer la pregunta y las opciones tipo test.
3. Seleccionar la opción correcta si está claramente justificada por el contenido del contexto, y justificar tu elección **citando explícitamente fragmentos relevantes del contexto**.
4. En caso de no estar claramente justificada por el contexto, intentar razonar desde el contenido del contexto una respuesta, indicando qué información se ha utilizado, pero NO UTILICES INFORMACIÓN EXTERNA
5. En caso de no encontrar información en el contexto relacionada con la pregunta, devolver "No existe información" aunque sepas la respuesta correcta.
"""

* Permite razonar solo si se citan y conectan fragmentos de forma explícita.

system_prompt = """
Eres un experto en Apache Spark y PySpark, y estás respondiendo preguntas tipo test usando exclusivamente la información del contexto proporcionado.

No puedes utilizar ningún conocimiento previo ni información externa al contexto.

Tu tarea es:
1. Leer el contexto cuidadosamente.
2. Leer la pregunta y las opciones tipo test.
3. Seleccionar la opción correcta si está claramente justificada por el contenido del contexto, y justificar tu elección citando fragmentos textuales exactos.
4. Si no hay una justificación directa, puedes razonar una respuesta **únicamente si se basa en información presente en el contexto**, indicando exactamente **qué fragmentos usas** y **cómo los conectas**. No está permitido inventar datos.
5. En caso de no encontrar ninguna información útil en el contexto, responde exactamente con: **"No existe información"**
"""

llm = OpenAI(
    model="gpt-4o-mini",
    temperature=0,
    system_prompt=system_prompt
)

Settings.llm = llm

def extract_predicted_letter(response_text):
    """
    Extrae la letra (A, B, C, D o E) de una respuesta generada en el formato:
    'Respuesta correcta: <LETRA>\nJustificación: <...>'
    """
    match = re.search(r"Respuesta\s+correcta\s*:\s*([A-E])", response_text, flags=re.IGNORECASE)
    if match:
        return match.group(1).upper()
    return None 

def ask_question_with_options(question_obj, 
                              #query_str_template, 
                              retriever = retriever, response_mode= "compact"):
    question_text = question_obj["question"]
    options_text = format_options(question_obj["options"])

    #qa_prompt = PromptTemplate(query_str_template)
    #Settings.text_qa_template = qa_prompt

    query_str = f"""
    {question_text}

    Opciones: 
    {options_text}
        
    IMPORTANTE:
        - Usa solo información del contexto.
        - No respondas con conocimientos externos.
        - Aunque sepas la respuesta correcta, si no existe esa información en el contexto devuelve "No existe información"
    
    Responde de la siguiente manera:
    "
    Respuesta correcta: <LETRA>
    Justificación: <JUSTIFICACIÓN>
    "
    """
    
    # Usar QueryBundle para pasar la información
    query_bundle = QueryBundle(
        query_str = query_str,
        custom_embedding_strs = [question_text, options_text]
    )

    response_synthesizer = get_response_synthesizer(response_mode=response_mode)
    
    query_engine = RetrieverQueryEngine(
        retriever=retriever,
        response_synthesizer=response_synthesizer
    )
    
    response = query_engine.query(query_bundle)
    predicted_answer = response.response.strip()
    
    return {
        "question_id": question_obj["question_id"],
        "question" : question_obj["question"],
        "options": format_options(question_obj["options"]),
        "predicted_answer": predicted_answer,
        "correct_answer": question_obj["correct_answer"],
        "is_correct": extract_predicted_letter(predicted_answer) == question_obj["correct_answer"],
        "response_obj": response
    }


# Preguntas que no razonaba:
ids = ["Test_1_17", "Test_2_28", "Test_2_31", "Test_1_41"]
for id in ids: 
    question =[q for q in questions if q["question_id"] == id][0]
    # result_random = ask_question_with_options(question, response_mode = "refine") # Sirve para Test_1_41 pero no para el resto. 
    result_random = ask_question_with_options(question, response_mode = "tree_summarize")
    display_question_result(result_random)
    print("="*20)

random_questions = random.sample(questions, k=3)

for question in random_questions: 
    
    print("Response mode: TREE SUMMARIZE")
    print("="*20)
    result_random = ask_question_with_options(question, response_mode = "tree_summarize")
    display_question_result(result_random)
    print("="*20)

    print("="*20)
    print("Response mode: REFINE")
    print("="*20)
    result_random = ask_question_with_options(question, response_mode = "refine") # Sirve para Test_1_41 pero no para el resto. 
    display_question_result(result_random)
    print("="*20)

# Test_2_57
question =[q for q in questions if q["question_id"] == "Test_2_57"][0]
test_result = ask_question_with_options(question, response_mode = "tree_summarize")
# Validation
display_question_result(test_result)

# Test_1_12 
question =[q for q in questions if q["question_id"] == "Test_1_12"][0]
test_result = ask_question_with_options(question, response_mode = "tree_summarize")
# Validation
display_question_result(test_result)
